Proper glBufferSubData
Finally figured out what the issue was by reading the docs (lolwut). Was getting an invalid operation error because offset+size was greater than GL_BUFFER_SIZE. Turns out when I was initializing the buffer with glBufferData I passed a size of 0 and thats why the error was being thrown. Now the buffer is initialized with the size of the mandelbrot struct (which should be the largest thing we pass it) and everything is great. Passing size=None to glBufferSubData makes it figure out the size based on the given data.