Bytes vs Bytearray?
All development to this point was done on a windows 10 PC with older nvidia drivers for an older 580 GTX card and bytearray() was perfectly fine. Now I'm on another win10 pc but with a newer nvidia card on the latest drivers and bytearray is no bueno. bytes() however does work fine, so yeah. Not sure why theres a big diff now.