Texture stuffs
Learning as I go but I think I got it mostly working now. Texture object is created and allocated as an RGBA32F. I really only want a single channel and I wanted unsigned ints but the internalformat specifiers are really specific and the error message are extremely generic. This works so far and but I think the data its creating has to be changed a bit to work. The current shader program seems to output good data. Just bookmarking my progress, its been a long day getting this to work. Thank god for the OpenGL superbible!

If this is working for sure the next step is to write the mandelbrot fragment shader as a compute shader instead.