buddhabrot test
Decided to try and test things without the compute shader by simply reading out the texture and writing it to a file. The results arent that great, it kind of looks like the anti-buddhabrot more than anything and it seems to be displaced down and to the left for some reason. I've made sure I'm reading the memory correctly so I think the problem is me not understanding the maths. And the planned compute shader will do all this on its own so most of this code will be deleted anyway.

After looking at the wiki description a bit more it looks like I should only be saving the values for points that do escape, whereas currently I mark all points the function travels through.