Buddhabrot?
I was trying to get really clever with dynamic arrays and stuff but GLSL doesn't play such games. In the end the only viable solution I came up with was just doing the calculations twice, once to see if it escaped, and once again to store the values the function went through. I might be able to do this with SSBOs but from my understanding if I used an SSBO with the fragment shader, every individual invocation of the shader will be modifying the same memory. What I need is for each invocation of the shader to have its own SSBO. Maybe I misunderstand how SSBOs work, I can make some test code to be sure.

For now it outputs something far closer to the buddhabrot images online. What bothers me still is this weird offset, and it appears smaller than it should be. But seriously, why is it offset low and left?? If I just write out the values given to fragColor then it comes out fine, no issues with offset. Might be something to do with my ortho->window coords code but thats tested fine before.

Also found a way to get better error output than the standard (and useless) 'invalid operation'. Whenever glUseProgram() causes a crash we check glGetProgramInfoLog() and get an actually helpful and detailed error message.